Updated FDModel to allow it to provide class mappings for dictionaries that aren't recognized by the FDDataClient's delegate.
A scenario could arise where subclass A of FDModel has a relationship
to subclass B of FDModel but subclass B has no way of identifying what
NSDictionary it can be created from.

This commit provides the ability for subclass A to provide the mapping
by implementing a method that takes in a NSDictionary and the remote
key path being parsed. The FDModel subclass can look at the dictionary
and based on its contents and the remote key path return the subclass
of FDModel that should be used to transform the dictionary into the
local model.