First phase of implementing FDModelProvider.
This is the start of a process that is designed to make FDModel a
standalone component that can be pulled out into its own repo.

The end goal is that FDModel does not need to only be used only through
FDDataClient. In theory someone could want to create instances of
FDModel from NSDictionary objects themselves or from data fetched from
a database.