Added debug helper methods to FDModel.
It has been difficulty debugging how a property for a FDModel is being
parsed by putting in breakpoints because the method that does the
transformation is recursive and so the breakpoints would be triggered
multiple times.

These helper methods allow you to implement them on the FDModel you are
trying to debug and put in breakpoints that will then only trigger on
that specific model.